Fixing CUDA implementation
There are still some problems with the update_acc_jrk
kernel.

+++ gravidy.cpp
@@ -12,11 +12,11 @@
 int n;
 int iterations;
 float total_mass;
-float int_time;
+double int_time;
 double ini_time, end_time;
 double init_time;
-float energy_ini, energy_end, energy_tmp;
-float ekin, epot;
+double energy_ini, energy_end, energy_tmp;
+double ekin, epot;
 
 // Struct vector to read the input file
 std::vector<particle> part;
@@ -29,16 +29,16 @@ double4 *d_p_r, *d_p_v;
 
 
 // Host pointers
-float *h_ekin, *h_epot;
-float *h_t, *h_dt;
+double *h_ekin, *h_epot;
+double *h_t, *h_dt;
 float   *h_m;
 int *h_move;
 double4 *h_r, *h_v, *h_a, *h_j;
 double4 *h_new_a, *h_new_j;
 
 // Device pointers
-float *d_ekin, *d_epot;
-float *d_t, *d_dt;
+double *d_ekin, *d_epot;
+double *d_t, *d_dt;
 float   *d_m;
 int *d_move;
 double4 *d_r, *d_v, *d_a, *d_j;
